+++ pkg/BUILD.bazel
@@ -41,8 +41,6 @@ ALL_TESTS = [
     "//pkg/ccl/changefeedccl:changefeedccl_test",
     "//pkg/ccl/multiregionccl:multiregionccl_test",
     "//pkg/ccl/multitenantccl/tenantcapabilitiesccl:tenantcapabilitiesccl_test",
-    "//pkg/ccl/multitenantccl/tenantcostserver/tenanttokenbucket:tenanttokenbucket_test",
-    "//pkg/ccl/multitenantccl/tenantcostserver:tenantcostserver_test",
     "//pkg/ccl/schemachangerccl/sctestbackupccl:sctestbackupccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupmrccl:sctestbackupmrccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl_test",
@@ -302,6 +300,8 @@ ALL_TESTS = [
     "//pkg/multitenant/tenantcapabilities:tenantcapabilities_test",
     "//pkg/multitenant/tenantcostclient:tenantcostclient_test",
     "//pkg/multitenant/tenantcostmodel:tenantcostmodel_test",
+    "//pkg/multitenant/tenantcostserver/tenanttokenbucket:tenanttokenbucket_test",
+    "//pkg/multitenant/tenantcostserver:tenantcostserver_test",
     "//pkg/obs/ash:ash_test",
     "//pkg/obs/clustermetrics/cmreader:cmreader_test",
     "//pkg/obs/clustermetrics/cmwatcher:cmwatcher_test",
@@ -1022,10 +1022,6 @@ GO_TARGETS = [
     "//pkg/ccl/multiregionccl:multiregionccl",
     "//pkg/ccl/multiregionccl:multiregionccl_test",
     "//pkg/ccl/multitenantccl/tenantcapabilitiesccl:tenantcapabilitiesccl_test",
-    "//pkg/ccl/multitenantccl/tenantcostserver/tenanttokenbucket:tenanttokenbucket",
-    "//pkg/ccl/multitenantccl/tenantcostserver/tenanttokenbucket:tenanttokenbucket_test",
-    "//pkg/ccl/multitenantccl/tenantcostserver:tenantcostserver",
-    "//pkg/ccl/multitenantccl/tenantcostserver:tenantcostserver_test",
     "//pkg/ccl/multitenantccl:multitenantccl",
     "//pkg/ccl/schemachangerccl/sctestbackupccl:sctestbackupccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupmrccl:sctestbackupmrccl_test",
@@ -1816,6 +1812,10 @@ GO_TARGETS = [
     "//pkg/multitenant/tenantcostclient:tenantcostclient_test",
     "//pkg/multitenant/tenantcostmodel:tenantcostmodel",
     "//pkg/multitenant/tenantcostmodel:tenantcostmodel_test",
+    "//pkg/multitenant/tenantcostserver/tenanttokenbucket:tenanttokenbucket",
+    "//pkg/multitenant/tenantcostserver/tenanttokenbucket:tenanttokenbucket_test",
+    "//pkg/multitenant/tenantcostserver:tenantcostserver",
+    "//pkg/multitenant/tenantcostserver:tenantcostserver_test",
     "//pkg/multitenant:multitenant",
     "//pkg/obs/ash:ash",
     "//pkg/obs/ash:ash_test",
