+++ pkg/multitenant/tenantcostclient/BUILD.bazel
@@ -55,7 +55,6 @@ go_test(
         "//pkg/base",
         "//pkg/ccl",
         "//pkg/ccl/changefeedccl",
-        "//pkg/ccl/multitenantccl/tenantcostserver",
         "//pkg/cloud",
         "//pkg/cloud/impl:cloudimpl",
         "//pkg/cloud/nodelocal",
@@ -69,6 +68,7 @@ go_test(
         "//pkg/multitenant",
         "//pkg/multitenant/multitenantio",
         "//pkg/multitenant/tenantcostmodel",
+        "//pkg/multitenant/tenantcostserver",
         "//pkg/roachpb",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
