+++ pkg/multitenant/tenantcostserver/BUILD.bazel
@@ -9,14 +9,14 @@ go_library(
         "system_table.go",
         "token_bucket.go",
     ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/multitenantccl/tenantcostserver",
+    importpath = "github.com/cockroachdb/cockroach/pkg/multitenant/tenantcostserver",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/base",
-        "//pkg/ccl/multitenantccl/tenantcostserver/tenanttokenbucket",
         "//pkg/kv",
         "//pkg/kv/kvpb",
         "//pkg/multitenant",
+        "//pkg/multitenant/tenantcostserver/tenanttokenbucket",
         "//pkg/roachpb",
         "//pkg/server",
         "//pkg/settings",
