+++ pkg/multitenant/tenantcostserver/tenanttokenbucket/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "tenanttokenbucket",
     srcs = ["tenant_token_bucket.go"],
-    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/multitenantccl/tenantcostserver/tenanttokenbucket",
+    importpath = "github.com/cockroachdb/cockroach/pkg/multitenant/tenantcostserver/tenanttokenbucket",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/kv/kvpb",
